+++ NEWS.md
@@ -1,3 +1,8 @@
+# dbw 0.2.0
+### Minor changes
+* Add ``vcov'' option, which indicates whether to estimate the variance.
+* Delete unused ``std'' function.
+
 # dbw 0.1.0
 ### Minor changes
 * Add references.